+++ fltk/src/menu.rs
@@ -177,7 +177,7 @@ impl MenuItem {
     /// Sets the label size of the menu item
     pub fn set_label_size(&mut self, sz: i32) {
         assert!(!self.was_deleted());
-        let sz = if sz < 1 { 1 } else { sz }; 
+        let sz = if sz < 1 { 1 } else { sz };
         unsafe { Fl_Menu_Item_set_label_size(self.inner, sz) }
     }